requirements.txt
requests
beautifulsoup4
python-dotenv
google-cloud-storage
